tests: added more component tests and some refactoring
