+++ package.json
@@ -61,7 +61,7 @@
     "devDependencies": {
         "@swellaby/eslint-config": "^1.1.6",
         "@swellaby/nyc-config": "^1.1.0",
-        "@swellaby/tslint-config": "^2.0.0",
+        "@swellaby/tslint-config": "^2.0.1",
         "@types/chai": "^4.1.7",
         "@types/mkdirp": "^0.5.2",
         "@types/mocha": "^5.2.5",
